Deleting the directory
Would like to delete this directory of the repository as some error in the folder is causing a lot of problem and the particular repository is not getting cloned. Please accept this request as I am not able to clone the repo and can't further contribute to the project repo. @Sulagna-Dutta-Roy